refactor(cli)!: fold `model-info` into `models info`
Remove the standalone top-level `harn model-info` command; its model-
metadata lookup now lives under the `models` noun as `harn models info`,
matching the provider/skill consolidations. `ModelInfoArgs` moves from
the provider module to the models module; `print_model_info` is reused
by the `models info` dispatch. No alias kept.
